Update README.md
Updated the script to use, instead of a Facebook Insights file, a "page_id". That makes it easier to use for everyone, but the problem right now is that the script doesn't have a start and end date (working on that). So expect half an hour running for a page with 3,000 posts.